Made animation speed independent of fps
... particle and object speed as well.

+++ Source Main 5.2/source/CSPetSystem.cpp
@@ -24,7 +24,7 @@
 
 extern bool g_PetEnableDuel;
 
-extern  float   WorldTime;
+extern  double   WorldTime;
 extern	char    TextList[50][100];
 extern	int     TextListColor[50];
 extern	int     TextBold[50];
